requirements.txt
blinker==1.4
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
six==1.11.0
SQLAlchemy==1.1.15
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
Werkzeug==0.12.2
Whoosh==2.7.4
